+++ sphinx/ext/duration.py
@@ -4,11 +4,10 @@ from __future__ import annotations
 
 import json
 import time
-from datetime import datetime, timedelta, timezone
 from itertools import islice
 from operator import itemgetter
-from pathlib import Path
-from typing import TYPE_CHECKING, cast
+from types import NoneType
+from typing import TYPE_CHECKING
 
 import sphinx
 from sphinx.domains import Domain
@@ -16,7 +15,8 @@ from sphinx.locale import __
 from sphinx.util import logging
 
 if TYPE_CHECKING:
-    from collections.abc import Set
+    from collections.abc import Collection, Set
+    from pathlib import Path
     from typing import TypedDict
 
     from docutils import nodes
@@ -26,14 +26,6 @@ if TYPE_CHECKING:
     class _DurationDomainData(TypedDict):
         reading_durations: dict[str, float]
 
-
-DEFAULT_OPTIONS = {
-    'duration_print_total': True,
-    'duration_print_slowest': True,
-    'duration_n_slowest': 5,
-    'duration_write_json': 'sphinx_reading_durations.json',
-    'duration_limit': None,
-}
 
 logger = logging.getLogger(__name__)
 
@@ -52,9 +44,9 @@ class DurationDomain(Domain):
 
     def warn_reading_duration(self, duration: float, duration_limit: float) -> None:
         logger.warning(
-            __('Reading duration %s exceeded the duration limit %s'),
-            _format_seconds(duration),
-            _format_seconds(duration_limit),
+            __('Reading duration %.3fs exceeded the duration limit %.3fs'),
+            duration,
+            duration_limit,
             type='duration',
             location=self.env.docname,
         )
@@ -95,10 +87,7 @@ def on_doctree_read(app: Sphinx, doctree: nodes.document) -> None:
     domain = app.env.domains['duration']
     domain.note_reading_duration(duration)
 
-    duration_limit = cast(
-        'float | None',
-        getattr(app.config, 'duration_limit', DEFAULT_OPTIONS['duration_limit']),
-    )
+    duration_limit: float | None = app.config.duration_limit
     if duration_limit is not None and duration > duration_limit:
         domain.warn_reading_duration(duration, duration_limit)
 
@@ -106,68 +95,57 @@ def on_doctree_read(app: Sphinx, doctree: nodes.document) -> None:
 def on_build_finished(app: Sphinx, error: Exception) -> None:
     """Display duration ranking on the current build."""
     domain = app.env.domains['duration']
-    reading_durations = domain.reading_durations
-    if not reading_durations:
+    if not domain.reading_durations:
         return
 
     # Get default options and update with user-specified values
-    options = DEFAULT_OPTIONS.copy()
-    for key in options:
-        options[key] = getattr(app.config, key, DEFAULT_OPTIONS[key])
+    if app.config.duration_print_total:
+        _print_total_duration(domain.reading_durations.values())
 
-    if options['duration_print_total']:
-        logger.info('')
-        logger.info(
-            __(
-                '====================== total reading duration =========================='
-            )
+    if app.config.duration_print_slowest:
+        _print_slowest_durations(
+            domain.reading_durations, app.config.duration_n_slowest
         )
 
-        n_files = len(reading_durations)
-        s = '' if n_files == 1 else 's'
+    if write_json := app.config.duration_write_json:
+        _write_json_durations(domain.reading_durations, app.outdir / write_json)
 
-        total = sum(reading_durations.values())
-        logger.info('Total time reading %d file%s:\n', n_files, s)
-        logger.info(_format_seconds(total, multiline=True))
 
-    if options['duration_print_slowest']:
-        sorted_durations = sorted(
-            reading_durations.items(), key=itemgetter(1), reverse=True
-        )
-        n_slowest = cast('int', options['duration_n_slowest'])
+def _print_total_duration(durations: Collection[float]) -> None:
+    logger.info('')
+    logger.info(
+        __('====================== total reading duration ==========================')
+    )
 
-        if n_slowest == 0:
-            n_slowest = len(sorted_durations)
-            fmt = ' '
-        else:
-            n_slowest = min(n_slowest, len(sorted_durations))
-            fmt = f' {n_slowest} '
+    n_files = len(durations)
+    s = 's' if n_files != 1 else ''
+    minutes, seconds = divmod(sum(durations), 60)
+    logger.info(
+        __('Total time reading %d file%s: %dm %.3fs'), n_files, s, minutes, seconds
+    )
 
-        logger.info('')
-        logger.info(
-            __(
-                f'====================== slowest{fmt}reading durations ======================='
-            )
-        )
 
-        for docname, d in islice(sorted_durations, n_slowest):
-            logger.info('%s %s', _format_seconds(d), docname)
+def _print_slowest_durations(durations: dict[str, float], n_slowest: int) -> None:
+    sorted_durations = sorted(durations.items(), key=itemgetter(1), reverse=True)
+    n_slowest = n_slowest or len(sorted_durations)
+    n_slowest = min(n_slowest, len(sorted_durations))
 
-        logger.info(__(''))
+    logger.info('')
+    logger.info('')
+    logger.info(
+        __('====================== slowest reading durations =======================')
+    )
+    for docname, duration in islice(sorted_durations, n_slowest):
+        logger.info(__('%.3fs %s'), duration, docname)
 
-    if write_json := options['duration_write_json']:
-        # Write to JSON
-        relpath = (
-            Path(write_json)
-            if isinstance(write_json, (Path, str))
-            else Path(cast('Path | str', DEFAULT_OPTIONS['duration_write_json']))
-        )
-        out_file = Path(app.builder.outdir) / relpath
-        # Make sure all parent dirs exist
-        for parent in out_file.parents[: len(relpath.parents) - 1]:
-            parent.mkdir(exist_ok=True)
-        with out_file.open('w', encoding='utf-8') as fid:
-            json.dump(reading_durations, fid, indent=4)
+    logger.info('')
+
+
+def _write_json_durations(durations: dict[str, float], out_file: Path) -> None:
+    durations = {k: round(v, 3) for k, v in durations.items()}
+    out_file.parent.mkdir(parents=True, exist_ok=True)
+    durations_json = json.dumps(durations, ensure_ascii=False, indent=4, sort_keys=True)
+    out_file.write_text(durations_json, encoding='utf-8')
 
 
 def setup(app: Sphinx) -> dict[str, bool | str]:
@@ -177,23 +155,21 @@ def setup(app: Sphinx) -> dict[str, bool | str]:
     app.connect('doctree-read', on_doctree_read)
     app.connect('build-finished', on_build_finished)
 
+    app.add_config_value('duration_print_total', True, '', types=frozenset({bool}))
+    app.add_config_value('duration_print_slowest', True, '', types=frozenset({bool}))
+    app.add_config_value('duration_n_slowest', 5, '', types=frozenset({int}))
+    app.add_config_value(
+        'duration_write_json',
+        'sphinx-reading-durations.json',
+        '',
+        types=frozenset({str, NoneType}),
+    )
+    app.add_config_value(
+        'duration_limit', None, '', types=frozenset({float, int, NoneType})
+    )
+
     return {
         'version': sphinx.__display_version__,
         'parallel_read_safe': True,
         'parallel_write_safe': True,
     }
-
-
-def _format_seconds(seconds: float, multiline: bool = False) -> str:
-    """Convert seconds to a formatted string."""
-    if not multiline:
-        return f'{seconds:.3f}s'
-    dt = datetime(1, 1, 1, tzinfo=timezone.utc) + timedelta(seconds=seconds)  # noqa: UP017
-    minutes = dt.hour * 60 + dt.minute
-    seconds = dt.second
-    milliseconds = round(dt.microsecond / 1000.0)
-    return (
-        f'minutes:      {minutes:>3}\n'
-        f'seconds:      {seconds:>3}\n'
-        f'milliseconds: {milliseconds:>3}'
-    )